+++ src/main/java/subway/dto/response/LineResponse.java
@@ -1,4 +1,4 @@
-package subway.dto;
+package subway.dto.response;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
