+++ src/main/java/subway/service/LineService.java
@@ -3,8 +3,8 @@ package subway.service;
 import org.springframework.stereotype.Service;
 import subway.dao.H2LineDao;
 import subway.domain.Line;
-import subway.dto.LineRequest;
-import subway.dto.LineResponse;
+import subway.dto.request.LineRequest;
+import subway.dto.response.LineResponse;
 
 import java.util.List;
 import java.util.stream.Collectors;
