+++ src/main/java/subway/service/SectionService.java
@@ -15,13 +15,13 @@ import subway.domain.Section;
 import subway.domain.Sections;
 import subway.domain.Station;
 import subway.dto.DtoMapper;
-import subway.dto.EndSectionRequest;
-import subway.dto.InitSectionRequest;
-import subway.dto.LineStationsResponse;
-import subway.dto.SectionDeleteRequest;
-import subway.dto.SectionLastDeleteRequest;
-import subway.dto.SectionRequest;
-import subway.dto.StationResponse;
+import subway.dto.request.EndSectionRequest;
+import subway.dto.request.InitSectionRequest;
+import subway.dto.response.LineStationsResponse;
+import subway.dto.request.SectionDeleteRequest;
+import subway.dto.request.SectionLastDeleteRequest;
+import subway.dto.request.SectionRequest;
+import subway.dto.response.StationResponse;
 
 import java.util.ArrayList;
 import java.util.List;
