+++ src/test/java/subway/integration/StationIntegrationTest.java
@@ -9,7 +9,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-import subway.dto.StationResponse;
+import subway.dto.response.StationResponse;
 
 import java.util.HashMap;
 import java.util.List;
